More patches
Working to ensure the differentials are being reported correctly

+++ server.c
@@ -180,7 +180,7 @@ int parse_client_command(RothagaClient *rc, RothagaClient *c)
 {
 	int l = 0;
 	char cmd[3];
-	float f = 0;
+	unsigned long int f = 0;
 
 	cmd[0] = 0;
 	cmd[1] = 0;
@@ -214,7 +214,7 @@ int parse_client_command(RothagaClient *rc, RothagaClient *c)
 	clock_gettime(CLOCK_MONOTONIC, &c->sndmes);
 	f = c->sndmes.tv_nsec - c->fstmes.tv_nsec;
 	
-	printf("Client %s message differential was %f\n",c->cliname,(f/1.0e6));
+	printf("Client %s message differential was %lu\n",c->cliname,(f/1.0e6));
 
 	if ((f / 1.0e6) < 1.0) c->karma--;
